pull out bulky css, rename bpToggleNavigation, refactor from React classes, create compound components for Navigation and Dropdown
